[TEST] - fix failing tests for repository(remove FactoryBean)
