+++ repository/src/org/pentaho/platform/repository2/unified/RepositoryAccessVoterManager.java
@@ -13,55 +13,56 @@
  * See the GNU General Public License for more details.
  *
  *
- * Copyright 2006 - 2014 Pentaho Corporation.  All rights reserved.
- *
- *   class instead repository.spring.xml changing:
- * 
- *  <bean id="repositoryAccessVoterManager" 
- *  class="org.pentaho.platform.repository2.unified.RepositoryAccessVoterManager">
- *    <constructor-arg ref="authorizationPolicy"/>
- *  </bean>
- * 
+ * Copyright 2006 - 2013 Pentaho Corporation.  All rights reserved.
  */
 
 package org.pentaho.platform.repository2.unified;
 
 import org.pentaho.platform.api.engine.IAuthorizationPolicy;
+import org.pentaho.platform.api.engine.IPentahoSession;
 import org.pentaho.platform.api.repository2.unified.IRepositoryAccessVoter;
-import org.springframework.beans.factory.FactoryBean;
+import org.pentaho.platform.api.repository2.unified.IRepositoryAccessVoterManager;
+import org.pentaho.platform.api.repository2.unified.RepositoryFile;
+import org.pentaho.platform.api.repository2.unified.RepositoryFileAcl;
+import org.pentaho.platform.api.repository2.unified.RepositoryFilePermission;
+import org.pentaho.platform.security.policy.rolebased.actions.AdministerSecurityAction;
+import org.springframework.util.Assert;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
 
-public class RepositoryAccessVoterManager implements FactoryBean {
+public class RepositoryAccessVoterManager implements IRepositoryAccessVoterManager {
 
-  private List<IRepositoryAccessVoter> voters;
   private IAuthorizationPolicy authorizationPolicy;
+  private List<IRepositoryAccessVoter> voters;
 
-  public RepositoryAccessVoterManager( final IAuthorizationPolicy authorizationPolicy,
-      final String repositoryAdminUsername ) {
-    this( Collections.<IRepositoryAccessVoter>emptyList(), authorizationPolicy, repositoryAdminUsername );
-  }
-
-
-  public RepositoryAccessVoterManager( final List<IRepositoryAccessVoter> voters,
-                                       final IAuthorizationPolicy authorizationPolicy, final String repositoryAdminUsername ) {
-    this.voters = voters;
+  public RepositoryAccessVoterManager(final IAuthorizationPolicy authorizationPolicy, final String repositoryAdminUsername) {
+    super();
     this.authorizationPolicy = authorizationPolicy;
   }
 
-  @Override
-  public Object getObject() throws Exception {
-    return new RepositoryAccessVoterManagerInst( voters, authorizationPolicy );
+  public RepositoryAccessVoterManager(final List<IRepositoryAccessVoter> voters,
+                                      final IAuthorizationPolicy authorizationPolicy, final String repositoryAdminUsername) {
+    this( authorizationPolicy, repositoryAdminUsername );
+    Assert.notNull( voters );
+    this.voters = new ArrayList<IRepositoryAccessVoter>();
+    this.voters.addAll( voters );
+  }
+
+  public void registerVoter( IRepositoryAccessVoter voter ) {
+    voters.add( voter );
   }
 
   @Override
-  public Class<?> getObjectType() {
-    return RepositoryAccessVoterManagerInst.class;
-  }
-
-  @Override
-  public boolean isSingleton() {
+  public boolean hasAccess( RepositoryFile file, RepositoryFilePermission operation,
+      RepositoryFileAcl repositoryFileAcl, IPentahoSession session ) {
+    if ( voters != null && !authorizationPolicy.isAllowed( AdministerSecurityAction.NAME ) ) {
+      for ( IRepositoryAccessVoter voter : voters ) {
+        if ( !voter.hasAccess( file, operation, repositoryFileAcl, session ) ) {
+          return false;
+        }
+      }
+    }
     return true;
   }
